controller/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ozma-controller"
version = "1.1.0"
description = "Ozma KVM switch controller — scenarios, audio, HID, streaming, AI agent"
readme = "README.md"
license = {text = "AGPL-3.0-only WITH OzmaPluginException"}
requires-python = ">=3.11"
authors = [{name = "Ozma Labs", email = "dev@ozmalabs.com"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "aiohttp>=3.9",
    "zeroconf>=0.80",
    "numpy>=1.24",
    "Pillow>=10.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
vision = ["ultralytics>=8.0", "torch>=2.0"]
midi = ["mido>=1.3", "python-rtmidi>=1.5"]
osc = ["python-osc>=1.8"]
streamdeck = ["elgato-streamdeck>=0.9", "Pillow>=10.0"]
terminal = ["chafa.py>=1.0"]
all = ["ozma-controller[vision,midi,osc,streamdeck,terminal]"]

[project.scripts]
ozma-controller = "main:main"
ozma-term-view = "term_view:main"

[project.urls]
Homepage = "https://ozma.dev"
Repository = "https://github.com/ozmalabs/ozma"
Documentation = "https://docs.ozma.dev"
